Factor out  server  config
