+++ internal/envstest/server.go
@@ -19,27 +19,23 @@ import (
 	"fmt"
 	"net"
 	"net/http"
-	"path/filepath"
 	"testing"
 	"time"
 
 	"github.com/google/exposure-notifications-server/pkg/keys"
 	"github.com/google/exposure-notifications-server/pkg/server"
 	"github.com/google/exposure-notifications-verification-server/internal/auth"
-	"github.com/google/exposure-notifications-verification-server/internal/project"
+	"github.com/google/exposure-notifications-verification-server/internal/envstest/testconfig"
 	"github.com/google/exposure-notifications-verification-server/internal/routes"
 	"github.com/google/exposure-notifications-verification-server/pkg/cache"
 	"github.com/google/exposure-notifications-verification-server/pkg/config"
 	"github.com/google/exposure-notifications-verification-server/pkg/controller"
 	"github.com/google/exposure-notifications-verification-server/pkg/database"
-	"github.com/google/exposure-notifications-verification-server/pkg/ratelimit"
 	"github.com/google/exposure-notifications-verification-server/pkg/rbac"
 
 	"github.com/gorilla/securecookie"
 	"github.com/gorilla/sessions"
-	"github.com/sethvargo/go-envconfig"
 	"github.com/sethvargo/go-limiter"
-	"github.com/sethvargo/go-limiter/memorystore"
 )
 
 const (
@@ -161,7 +157,7 @@ func NewServer(tb testing.TB, testDatabaseInstance *database.TestInstance) *Test
 	}
 
 	// Create the config and requirements.
-	response := newServerConfig(tb, testDatabaseInstance)
+	response := testconfig.NewServerConfig(tb, testDatabaseInstance)
 
 	ctx := context.Background()
 
@@ -210,129 +206,4 @@ func NewServer(tb testing.TB, testDatabaseInstance *database.TestInstance) *Test
 		RateLimiter:  response.RateLimiter,
 		Server:       srv,
 	}
-}
-
-// serverConfigResponse is the response from creating a server config.
-type serverConfigResponse struct {
-	AuthProvider auth.Provider
-	Config       *config.ServerConfig
-	Database     *database.Database
-	Cacher       cache.Cacher
-	KeyManager   keys.KeyManager
-	RateLimiter  limiter.Store
-}
-
-// newServerConfig creates a new server configuration. It creates all the keys,
-// databases, and cacher, but does not actually start the server. All cleanup is
-// scheduled by t.Cleanup.
-func newServerConfig(tb testing.TB, testDatabaseInstance *database.TestInstance) *serverConfigResponse {
-	tb.Helper()
-
-	if testing.Short() {
-		tb.Skip()
-	}
-
-	// Create the auth provider
-	authProvider, err := auth.NewLocal(context.Background())
-	if err != nil {
-		tb.Fatal(err)
-	}
-
-	// Create the cacher.
-	cacher, err := cache.NewInMemory(nil)
-	if err != nil {
-		tb.Fatal(err)
-	}
-	tb.Cleanup(func() {
-		if err := cacher.Close(); err != nil {
-			tb.Fatal(err)
-		}
-	})
-
-	// Create the database.
-	db, dbConfig := testDatabaseInstance.NewDatabase(tb, cacher)
-
-	// Create the key manager.
-	keyManager := keys.TestKeyManager(tb)
-
-	// Create the rate limiter.
-	limiterStore, err := memorystore.New(&memorystore.Config{
-		Tokens:   30,
-		Interval: time.Second,
-	})
-	if err != nil {
-		tb.Fatal(err)
-	}
-	tb.Cleanup(func() {
-		ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-
-		if err := limiterStore.Close(ctx); err != nil {
-			tb.Fatal(err)
-		}
-	})
-
-	// Create the config.
-	cfg := &config.ServerConfig{
-		AssetsPath:  ServerAssetsPath(),
-		LocalesPath: LocalesPath(),
-		Cache: cache.Config{
-			Type:    cache.TypeInMemory,
-			HMACKey: RandomBytes(tb, 64),
-		},
-		Database: *dbConfig,
-		// Firebase is not used for browser tests.
-		Firebase: config.FirebaseConfig{
-			APIKey:          "test",
-			AuthDomain:      "test.firebaseapp.com",
-			DatabaseURL:     "https://test.firebaseio.com",
-			ProjectID:       "test",
-			StorageBucket:   "test.appspot.com",
-			MessageSenderID: "test",
-			AppID:           "1:test:web:test",
-			MeasurementID:   "G-TEST",
-		},
-		CookieKeys:  config.Base64ByteSlice{RandomBytes(tb, 64), RandomBytes(tb, 32)},
-		CSRFAuthKey: RandomBytes(tb, 32),
-		CertificateSigning: config.CertificateSigningConfig{
-			CertificateSigningKey: "UPDATE_ME",
-			Keys: keys.Config{
-				KeyManagerType: keys.KeyManagerTypeFilesystem,
-				FilesystemRoot: filepath.Join(project.Root(), "local", "test", RandomString(tb, 8)),
-			},
-		},
-		RateLimit: ratelimit.Config{
-			Type:    ratelimit.RateLimiterTypeMemory,
-			HMACKey: RandomBytes(tb, 64),
-		},
-
-		// DevMode has to be enabled for tests. Otherwise the cookies fail.
-		DevMode: true,
-	}
-
-	// Process the config - this simulates production setups and also ensures we
-	// get the defaults for any unset values.
-	emptyLookuper := envconfig.MapLookuper(nil)
-	if err := config.ProcessWith(context.Background(), cfg, emptyLookuper); err != nil {
-		tb.Fatal(err)
-	}
-
-	return &serverConfigResponse{
-		AuthProvider: authProvider,
-		Config:       cfg,
-		Database:     db,
-		Cacher:       cacher,
-		KeyManager:   keyManager,
-		RateLimiter:  limiterStore,
-	}
-}
-
-// ServerAssetsPath returns the path to the UI server assets.
-func ServerAssetsPath() string {
-	return filepath.Join(project.Root(), "cmd", "server", "assets")
-}
-
-// LocalesPath returns the path to the i18n locales.
-func LocalesPath() string {
-	return filepath.Join(project.Root(), "internal", "i18n", "locales")
 }
