+++ internal/envstest/testconfig/random.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package envstest
+package testconfig
 
 import (
 	"crypto/rand"
